Moved IPC stream sink into stream_async file.
