+++ src/io/ipc/write/mod.rs
@@ -18,9 +18,6 @@ mod common_async;
 #[cfg(feature = "io_ipc_write_async")]
 #[cfg_attr(docsrs, doc(cfg(feature = "io_ipc_write_async")))]
 pub mod stream_async;
-#[cfg(feature = "io_ipc_write_async")]
-#[cfg_attr(docsrs, doc(cfg(feature = "io_ipc_write_async")))]
-pub mod sink;
 
 #[cfg(feature = "io_ipc_write_async")]
 #[cfg_attr(docsrs, doc(cfg(feature = "io_ipc_write_async")))]
